Update sonoff-wifi-switch.groovy
This patch is adding IP address to config panel, because it happened to me that after 2 weeks of using this DTH the IP changed to 0.0.0.0 by itself. It just happened. With this patch, You can set back IP in Device settings in SmartThings App->Things.
Otherwise, You would have to add this Device again, and then binding with many Routines and SmartApps from the old Device that lost it's IP would be lost.